requirements.txt
streamlit==1.43.2
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.2
